# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Durian HiTech",
        "title": "Durian Covid-19 Golang Backend",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/": {
            "get": {
                "description": "测试 Index 页",
                "tags": [
                    "测试"
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"message\": \"gcp\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/news/detail": {
            "post": {
                "description": "查看单条新闻",
                "tags": [
                    "新闻"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "新闻ID",
                        "name": "news_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true, \"message\":\"查询成功\",\"data\":\"该条新闻的详细信息\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"success\":true, \"message\":\"查询失败，新闻ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/news/list_all_news": {
            "get": {
                "description": "获取所有新闻，返回列表",
                "tags": [
                    "新闻"
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true, \"message\":\"查询成功\",\"data\":\"所有新闻\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notice/create_comment": {
            "post": {
                "description": "创建一条评论",
                "tags": [
                    "防控知识板块"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户类型",
                        "name": "user_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "问题ID",
                        "name": "question_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "评论内容",
                        "name": "comment_content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"message\": \"用户评论成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"message\": \"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"message\": \"问题ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"success\": false, \"message\": \"评论失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notice/create_question": {
            "post": {
                "description": "创建一个问题",
                "tags": [
                    "防控知识板块"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "提问标题",
                        "name": "question_title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "提问内容",
                        "name": "question_content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"message\": \"用户提问成功\", \"detail\": 提问的全部信息}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"message\": \"数据库error, 一些其他错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"success\": false, \"message\": \"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notice/list_all_comments": {
            "post": {
                "description": "列出某个问题的全部评论",
                "tags": [
                    "防控知识板块"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "问题ID",
                        "name": "question_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"message\": \"查看成功\", \"data\": \"某问题的所有评论\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"success\": false, \"message\": \"问题ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sub/del_sub": {
            "post": {
                "description": "删除订阅",
                "tags": [
                    "订阅城市"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "订阅ID",
                        "name": "subscription_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true, \"message\":\"删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"message\": \"数据库error, 一些其他错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"success\": false, \"message\": \"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sub/list_all_subs": {
            "post": {
                "description": "获取订阅列表",
                "tags": [
                    "订阅城市"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true, \"message\":\"查询成功\",\"data\":\"user的所有订阅\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"success\": false, \"message\": \"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sub/subscribe": {
            "post": {
                "description": "订阅城市疫情信息",
                "tags": [
                    "订阅城市"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "城市名字",
                        "name": "city_name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true, \"message\":\"订阅成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"success\": false, \"message\": \"数据库error, 一些其他错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"success\": false, \"message\": \"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "post": {
                "description": "查看用户个人信息",
                "tags": [
                    "用户管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"message\": \"查看用户信息成功\", \"data\": \"model.User的所有信息\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"success\": false, \"message\": \"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "登录",
                "tags": [
                    "用户管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"message\": \"登录成功\", \"detail\": user的信息}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"message\": \"密码错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"success\": false, \"message\": \"没有该用户\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/modify": {
            "post": {
                "description": "修改用户信息（支持修改用户名和密码）",
                "tags": [
                    "用户管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "原密码",
                        "name": "password_old",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新密码",
                        "name": "password_new",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"message\": \"修改成功\", \"data\": \"model.User的所有信息\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"message\": \"原密码输入错误\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"success\": false, \"message\": \"用户ID不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "注册",
                "tags": [
                    "用户管理"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户类型（0: 普通用户，1: 认证机构用户）",
                        "name": "user_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "认证机构名",
                        "name": "affiliation",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\": true, \"message\": \"用户创建成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"success\": false, \"message\": \"用户已存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}